+++ workbox-config.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 module.exports = {
     globDirectory: `build/`,
-    globPatterns: [`**/*.{css,html,ico,js,json,mem,png,zip}`],
+    globPatterns: [`**/*.{css,html,ico,js,json,mem,png,wasm,zip}`],
     maximumFileSizeToCacheInBytes: `8000000`,
     swDest: `build/sw.js`,
     swSrc: `src/sw.js`,